apparently working batch_script.py, untested dispatch.py script, no compile or json compile script yet

+++ dispatch_script.py
@@ -1,41 +1,52 @@
 #!/usr/bin/python
+"""Dispatch parallel job to compute dependency matrix.
+
+EXAMPLE USE:
+python $HOME/recomb2013_workbench/dependency_matrix/dispatch_script.py fname1=/fs/lustre/osu6683/recomb2013_gse15745/GSE15745_GPL6104.data.aligned.pkl fname2=/fs/lustre/osu6683/recomb2013_gse15745/GSE15745_GPL8490.data.aligned.pkl computers=[\"Dcor\",\"PCC\"] outdir=/fs/lustre/osu6683/recomb2013_gse15745/dispatch_test dry=True
+"""
 import datetime
 import random
 from __init__ import *
 from matrix_io import *
 from qsub import *
+import json
 import shutil
+import sys
 
 
 bname = os.path.basename
 K = 100000
 def load_cp_and_make_bin(fname, outdir):
-  d = load(fname); M, ftype = d["M"], d["ftype"]
-  if ftype != ".pkl":
-    fname_new = os.path.join(outdir, bname(fname.rpartition('.')[2])+".pkl")
+  d = load(fname)
+  M, ftype = d["M"], d["ftype"]
+  if ftype != "pkl":
+    fname_new = os.path.join(outdir, bname(fname.rpartition('.')[0])+".pkl")
     save(M, fname_new)
     print "Saved binary copy of matrix %s as %s" % (fname, fname_new)
   else:
     fname_new = os.path.join(outdir, bname(fname))
-    if os.path.abspath(fname) != os.path.abspath(fname_new):
+    if os.path.abspath(fname) != os.path.abspath(fname_new) and not os.path.exists(fname_new):
+      print "Matrix is not in out directory %s. Copying to %s" % (outdir, fname_new)
       shutil.copy(fname, fname_new)
   return (M, fname_new)
 
 
-def main(fname=None, fname1=None, fname2=None, computers=None, outdir=None, n_nodes=1, n_ppn=1, hours=8, compute_options=None):
+def main(fname=None, fname1=None, fname2=None, computers=None, outdir=None, n_nodes=1, n_ppn=1, hours=8, compute_options=None, dry=False):
 
   assert bool(fname) != bool(fname1 and fname2)
   if computers is None:
     computers = COMPUTERS.keys()
   else:
-    assert not set(computers) - set(COMPUTERS.key())
+    if isinstance(computers, basestring):
+      computers = json.loads(computers)
+    assert not set(computers) - set(COMPUTERS.keys())
   if not os.path.exists(outdir):
     make_dir(outdir)
     print "Created outdir %s" % outdir
   assert n_nodes >= 1
   assert n_ppn >= 1 and n_ppn <= 12
   assert hours >= 1 and hours <= 99
-  os.path.cwd(outdir)
+  os.chdir(outdir)
   compiled_dir = os.path.join(outdir, "compiled")
 
   if fname:
@@ -66,13 +77,24 @@ def main(fname=None, fname1=None, fname2=None, computers=None, outdir=None, n_no
       Q = Qsub(jobname=jobname, n_nodes=n_nodes, n_ppn=n_ppn, hours=hours, work_dir=outdir)
       cmds = shells_dispatch_dual(fname_work1, fname_work2, np.size(M2,0), comp_name, compute_options)
       Q.add_parallel(cmds)
-      
-    Q.add(shell_compile(compiled_dir, comp_dir, mtype))
-    pid = Q.submit()
+
+    cmd = shell_compile(compiled_dir, comp_dir, mtype)
+    Q.add(cmd)
+    pid = Q.submit(dry)
+    print Q.script()
+    print "Submitted dispatch, job ID: %s" % pid
     pids.append(pid)
 
   # After all jobs have completed, compile .json job and send email on completion.
   Q = Qsub(jobname="%s_JSONCompile"%fname, work_dir=outdir, email=True, after_jobids=pids)
   Q.add(shell_jsonindex(compiled_dir, comp_dir, mtype))
-  return Q.submit()
+  print Q.script()
+  pid = Q.submit(dry)
+  print "Final PID: %s" % pid
+  return pid
 
+
+if __name__ == "__main__":
+  kwds = dict([s.split('=') for s in sys.argv[1:]])
+  print kwds
+  main(**kwds)
